Separate enums for model name per LLM
