+++ src/main/java/com/mulesoft/connectors/internal/models/OllamaModelName.java
@@ -1,0 +1,16 @@
+package com.mulesoft.connectors.internal.models;
+
+enum OllamaModelName {
+    MISTRAL("mistral"), PHI3("phi3"), ORCA_MINI("orca-mini"), LLAMA2("llama2"), CODE_LLAMA("codellama"), TINY_LLAMA("tinyllama");
+
+    private final String value;
+
+    OllamaModelName(String value) {
+        this.value = value;
+    }
+
+    @Override
+    public String toString() {
+        return this.value;
+    }
+}
